+++ src/test/java/domain/LadderTest.java
@@ -1,48 +1,59 @@
 package domain;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import org.junit.jupiter.params.provider.ValueSource;
 
-import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 
-class LadderTest {
+import static java.lang.Boolean.FALSE;
+import static java.lang.Boolean.TRUE;
+import static org.assertj.core.api.Assertions.assertThat;
 
-    @DisplayName("사다리의 높이와 사람 수만큼 라인을 생성한다")
-    @ParameterizedTest
-    @CsvSource({
-            "2, 1",
-            "2, 10",
-            "3, 5",
-    })
-    void ladder_test(int personCount, int maxHeight) {
-        List<String> list = new ArrayList<>();
-        for (int i = 0; i < personCount; i++) {
-            list.add("hi" + i);
-        }
-        Players players = new Players(list);
-        Results results = new Results(list);
-        Lines lines = new RandomLadderGenerator()
-                .generate(players, maxHeight);
-        Ladder ladder = new Ladder(players, results, lines);
-        assertThat(ladder.getWidth()).isEqualTo(personCount - 1);
-        assertThat(ladder.getHeight()).isEqualTo(maxHeight);
+public class LadderTest {
+
+    Ladder ladder;
+
+    @BeforeEach
+    void init() {
+        Players players = new Players(List.of("pobi", "honux", "crong", "jk"));
+        Results results = new Results(List.of("꽝", "5000", "꽝", "3000"));
+        List<Line> customizedLines = List.of(
+                new Line(List.of(TRUE, FALSE, TRUE)),
+                new Line(List.of(FALSE, TRUE, FALSE)),
+                new Line(List.of(TRUE, FALSE, FALSE)),
+                new Line(List.of(FALSE, TRUE, FALSE)),
+                new Line(List.of(TRUE, FALSE, TRUE))
+        );
+        ladder = new Ladder(players, results, new Lines(customizedLines));
     }
 
-    @DisplayName("사다리의 높이가 1 이상 10 이하가 아니면 예외가 발생한다.")
-    @ParameterizedTest(name = "{index} : 현재 사다리 높이 = {0}")
-    @ValueSource(ints = {0, 11})
-    void ladder_height_test(int height) {
-        Players players = new Players(List.of("a", "b", "c"));
-        Results results = new Results(List.of("1", "2", "3"));
-        assertThatThrownBy(() -> new RandomLadderGenerator()
-                .generate(players, height))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("사다리 높이는 1 이상 10 이하여야 합니다.");
+    @DisplayName("단일 사용자의 결과 출력")
+    @ParameterizedTest
+    @CsvSource({
+            "pobi,꽝",
+            "honux,3000",
+            "crong,꽝",
+            "jk,5000",
+    })
+    void single_result_test(String name, String result) {
+        assertThat(ladder.calculateResult(new Player(name)))
+                .isEqualTo(new Result(result));
+    }
+
+    @DisplayName("모든 사용자의 결과 출력")
+    @Test
+    void all_result_test() {
+        Map<Player, Result> data = Map.ofEntries(
+                Map.entry(new Player("pobi"), new Result("꽝")),
+                Map.entry(new Player("honux"), new Result("3000")),
+                Map.entry(new Player("crong"), new Result("꽝")),
+                Map.entry(new Player("jk"), new Result("5000"))
+        );
+        CalculatedResults resultMap = ladder.getTotalResults();
+        assertThat(resultMap.getResultMap()).isEqualTo(data);
     }
 }
